impl (minor): In InterestFilterTable, remove unnecessary cast of int to int.

+++ src/net/named_data/jndn/impl/InterestFilterTable.java
@@ -135,7 +135,7 @@ public class InterestFilterTable {
     int count = 0;
     // Go backwards through the list so we can remove entries.
     // Remove all entries even though interestFilterId should be unique.
-    for (int i = (int)table_.size() - 1; i >= 0; --i) {
+    for (int i = table_.size() - 1; i >= 0; --i) {
       if (((Entry)table_.get(i)).getInterestFilterId() == interestFilterId) {
         ++count;
         table_.remove(i);
